api/requirements.txt
fastapi == 0.89.1
mysqlclient == 2.1.1
uvicorn[standard] == 0.20.0
Pillow == 9.3.0